Skip if existing file found

+++ javascript/sfc-licensees.js
@@ -54,18 +54,20 @@ const scrapSFCLicensees = async pageNum => {
 	return result.join("\n");
 };
 
-const pageNumList = [1440, 1441, 1445, 49572, 21177, 45382, 14215, 77546]; // set required page numbers to scrap
-
-// const pageNumList = [77546]; // set required page numbers to scrap
+const pageNumList = [1440, 49572, 21177, 45382, 14215, 77546]; // set required page numbers to scrap
 
 pageNumList.forEach(pageNum => {
-	scrapSFCLicensees(pageNum).then(result => {
+	if (fs.existsSync(`./sfc-licensees-data/sfc-licensees-${pageNum}.csv`)) {
+		console.log(`Skipping ${pageNum}, existing file found!`);
+		return;
+	}
+	return scrapSFCLicensees(pageNum).then(result => {
 		fs.writeFile(
 			`./sfc-licensees-data/sfc-licensees-${pageNum}.csv`,
 			result,
 			function (err) {
 				if (err) throw err;
-				console.log("Done!");
+				console.log(`Done scrapping ${pageNum}!`);
 			}
 		);
 	});
